:recycle:(trainer): extract feature-builder from market-data-buffer.ts
- Move buildFeatures and orderBookAverages to new feature-builder.ts
- Extract shared types (SymbolState, FEATURE_DIM, TradingSymbol) to
  market-data-types.ts to break circular import
- market-data-buffer.ts reduced from 306 to ~160 lines

Refs #176
